Fixing compiler warning about sign conversion.

+++ Pickers/ActionSheetDatePicker.m
@@ -88,7 +88,7 @@
     NSInteger index = button.tag;
     NSAssert((index >= 0 && (NSUInteger)index < self.customButtons.count), @"Bad custom button tag: %ld, custom button count: %lu", (long)index, (unsigned long)self.customButtons.count);
     NSAssert([self.pickerView respondsToSelector:@selector(setDate:animated:)], @"Bad pickerView for ActionSheetDatePicker, doesn't respond to setDate:animated:");
-    NSDictionary *buttonDetails = [self.customButtons objectAtIndex:index];
+    NSDictionary *buttonDetails = [self.customButtons objectAtIndex:(NSUInteger)index];
     NSDate *itemValue = [buttonDetails objectForKey:@"buttonValue"];
     UIDatePicker *picker = (UIDatePicker *)self.pickerView;    
     [picker setDate:itemValue animated:YES];
